sami+chedli+mariem+yassine+oussama revision

+++ src/main/java/com/repositories/ProfileRepository.java
@@ -12,14 +12,14 @@ import org.springframework.stereotype.Repository;
 import com.entities.Profile;
 
 @Repository
-public interface ProfileRepository extends CrudRepository<Profile, Integer> {
+public interface ProfileRepository extends CrudRepository<Profile, Long> {
 
 	@Query(value = "SELECT * FROM Profile  WHERE email = ?1", nativeQuery = true)
 	public Profile findByEmail(String email);
 
 	@Modifying
 	@Query(value = "UPDATE Profile SET enabled = true WHERE idProfile = ?1", nativeQuery = true)
-	public void enable(Integer id);
+	public void enable(Long id);
 
 	@Query(value = "SELECT * FROM Profile  WHERE verification_Code = ?1", nativeQuery = true)
 	public Profile findByVerificationCode(String code);
